client/requirements.txt
gradio==4.44.1
requests==2.32.3
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5